Dev back gmail

+++ backend/models/userratings.py
@@ -5,7 +5,7 @@ from typing import List
 class UserRating(BaseModel):
     film_id: int
     user_id: int
-    rating: int  # Число от 1 до 5
+    rating: float  # Число от 0.5 до 5
 
 # Хранилище оценок (для демонстрации)
 ratings: List[UserRating] = []
